+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Change Log
+# Label Craft
+## Change Log
+
+[comment]: <>  (2024-12-10 : YEAR-MONTH-DAY)
+
+## [1.0.11] - 2024-12-11
++ add Shuffle2 integration
 
 ## [1.0.10] - 2024-12-10
 + add full black and full white to Read/ Shuffle class